refactor: :recycle: model fix
created model for all movies and series

none
