+++ app/(tabs)/index.tsx
@@ -11,11 +11,11 @@ import React from "react";
 import { SafeAreaView } from "react-native";
 import { useQuery } from "@tanstack/react-query";
 import { getDiscoverSeries } from "../service/getDiscoverSeries";
-import { Result } from "../model/nowPlayingMovies";
+import { Result } from "../model/movieModel";
 import { LinearGradient } from "expo-linear-gradient";
 import MoviePoster from "../components/moviePoster";
 import SeriesPoster from "../components/seriesPoster";
-import { DiscoverSeries } from "../model/discoverSeries";
+import { SeriesModel } from "../model/seriesModel";
 import PopularMoviePoster from "../components/popularMoviePoster";
 import { getPopularMovies } from "../service/getPopularMovies";
 import PopularSeriesPoster from "../components/popularSeriesPoster";
