--- app/model/discoverSeries.ts
@@ -1,24 +1,0 @@
-export interface DiscoverSeries {
-    page: number
-    results: Result[]
-    total_pages: number
-    total_results: number
-  }
-  
-  export interface Result {
-    adult: boolean
-    backdrop_path?: string
-    genre_ids: number[]
-    id: number
-    origin_country: string[]
-    original_language: string
-    original_name: string
-    overview: string
-    popularity: number
-    poster_path: string
-    first_air_date: string
-    name: string
-    vote_average: number
-    vote_count: number
-  }
-
