+++ app/model/seriesModel.ts
@@ -1,4 +1,4 @@
-export interface NowPlayingMovie {
+export interface SeriesModel {
     page: number
     results: Result[]
     total_pages: number
@@ -7,17 +7,18 @@ export interface NowPlayingMovie {
   
   export interface Result {
     adult: boolean
-    backdrop_path: string
+    backdrop_path?: string
     genre_ids: number[]
     id: number
+    origin_country: string[]
     original_language: string
-    original_title: string
+    original_name: string
     overview: string
     popularity: number
     poster_path: string
-    release_date: string
-    title: string
-    video: boolean
+    first_air_date: string
+    name: string
     vote_average: number
     vote_count: number
   }
+
